# BEAST-Server/res/content/api-docs/OpenAPI.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "D-FARM API",
    "description" : "D-FARM API",
    "contact" : {
      "name" : "CodeInspect Support Team",
      "url" : "https://codeinspect.de",
      "email" : "helpdesk@codeinspect.de"
    },
    "version" : "0.2"
  },
  "security" : [ {
    "APIKey" : [ ]
  } ],
...
  "paths" : {
...
    "/api/devices/{devid}/network/ports/onhost/{port}" : {
      "put" : {
        "tags" : [ "Port forwardings" ],
        "summary" : "Creates a port forwarding on host",
        "description" : "Forward a local port (not a device port). This is necessary to debug using adb, which only allows debug clients to attach from localhost.",
        "operationId" : "createLocalPortForwarding",
        "parameters" : [ {
          "name" : "devid",
          "in" : "path",
          "description" : "The id of device (ignored, used only to determine which controller is responsible)",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "port",
          "in" : "path",
          "description" : "The port on the host",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          },
          "example" : 4000
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "integer",
                  "format" : "int32"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/APIException"
                }
              }
            }
          },
          "400" : {
            "description" : "Wrong input",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/APIException"
                }
              }
            }
          },
          "403" : {
            "description" : "Access denied",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/APIException"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "Port forwardings" ],
        "summary" : "Delete a port forwarding on host",
        "description" : "Stop forwarding a local port.",
        "operationId" : "deleteLocalPortForwarding",
        "parameters" : [ {
          "name" : "devid",
          "in" : "path",
          "description" : "The id of device (ignored, used only to determine which controller is responsible)",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "port",
          "in" : "path",
          "description" : "The port on the host",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          },
          "example" : 4000
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "integer",
                  "format" : "int32"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/APIException"
                }
              }
            }
          },
          "400" : {
            "description" : "Wrong input",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/APIException"
                }
              }
            }
          },
          "403" : {
            "description" : "Access denied",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/APIException"
                }
              }
            }
          }
        }
      }
    },
    "/api/devices/{devid}/network/ports/ondevice" : {
      "put" : {
        "tags" : [ "Port forwardings" ],
        "summary" : "Creates a port forwarding",
        "description" : "Creates a port forwarding",
        "operationId" : "createPortForwarding",
        "parameters" : [ {
          "name" : "devid",
          "in" : "path",
          "description" : "The id of device",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "portOnDevice",
          "in" : "query",
          "description" : "The port on the device",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          },
          "example" : 4000
        }, {
          "name" : "direction",
          "in" : "query",
          "description" : "The direction of the port forwarding",
          "required" : true,
          "schema" : {
            "type" : "string",
            "enum" : [ "DEVICE_IS_SERVER", "DEVICE_IS_CLIENT" ]
          },
          "example" : true
        }, {
          "name" : "protocol",
          "in" : "query",
          "description" : "The protocol.",
          "required" : true,
          "schema" : {
            "type" : "string",
            "enum" : [ "TCP", "JDWP" ]
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PortForwarding"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/APIException"
                }
              }
            }
          },
          "400" : {
            "description" : "Wrong input",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/APIException"
                }
              }
            }
          },
          "403" : {
            "description" : "Access denied",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/APIException"
                }
              }
            }
          }
        }
      }
    },
    "/api/devices/{devid}/network/ports" : {
      "get" : {
        "tags" : [ "Port forwardings" ],
        "summary" : "Lists port forwardings",
        "description" : "Lists port forwardings",
        "operationId" : "getPorts",
        "parameters" : [ {
          "name" : "devid",
          "in" : "path",
          "description" : "The id of device",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/PortForwarding"
                  }
                }
              }
            }
          },
          "500" : {
            "description" : "Internal error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/APIException"
                }
              }
            }
          },
          "400" : {
            "description" : "Wrong input",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/APIException"
                }
              }
            }
          },
          "403" : {
            "description" : "Access denied",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/APIException"
                }
              }
            }
          }
        }
      }
    },
    "/api/devices/{devid}/network/ports/ondevice/{id}" : {
      "delete" : {
        "tags" : [ "Port forwardings" ],
        "summary" : "Deletes a port forwarding",
        "description" : "Deletes a port forwarding",
        "operationId" : "removeForwarding",
        "parameters" : [ {
          "name" : "devid",
          "in" : "path",
          "description" : "The id of device",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "id",
          "in" : "path",
          "description" : "The id of the forwarding",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          },
          "500" : {
            "description" : "Internal error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/APIException"
                }
              }
            }
          },
          "400" : {
            "description" : "Wrong input",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/APIException"
                }
              }
            }
          },
          "403" : {
            "description" : "Access denied",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/APIException"
                }
              }
            }
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
...
      "APIException" : {
        "required" : [ "description", "title" ],
        "type" : "object",
        "properties" : {
          "title" : {
            "type" : "string",
            "description" : "A short description of the error that occured",
            "example" : "Internal Error"
          },
          "description" : {
            "type" : "string",
            "description" : "A more detailed description of the error that occured",
            "example" : "Out of storage space"
          },
          "code" : {
            "type" : "integer",
            "format" : "int32"
          }
        },
        "description" : "An exception has occurred while processing the request."
      },
...
      "PortForwarding" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int32"
          },
          "portOnHostMachine" : {
            "type" : "integer",
            "format" : "int32"
          },
          "portOnDevice" : {
            "type" : "integer",
            "format" : "int32"
          },
          "direction" : {
            "type" : "string",
            "enum" : [ "DEVICE_IS_SERVER", "DEVICE_IS_CLIENT" ]
          },
          "protocolOnDevice" : {
            "type" : "string",
            "enum" : [ "TCP", "JDWP" ]
          }
        }
      }
    },
    "securitySchemes" : {
      "APIKey" : {
        "type" : "apiKey",
        "description" : "Your API key",
        "name" : "APIKey",
        "in" : "header"
      }
    }
  }
}